refact: #no-issue refact ShareSettings, PostEditItem, QuestionItem  with typescript, update eslint
