--- frontend/src/components/_common/share-settings/ShareSettings.jsx
@@ -1,182 +1,0 @@
-import React, { useState, useEffect } from 'react';
-import { FormControlLabel, Button, Checkbox } from '@material-ui/core';
-import Typography from '@material-ui/core/Typography';
-import { useDispatch } from 'react-redux';
-import { useLocation, useHistory } from 'react-router-dom';
-import Dialog from '@material-ui/core/Dialog';
-import DialogActions from '@material-ui/core/DialogActions';
-import DialogContent from '@material-ui/core/DialogContent';
-import DialogContentText from '@material-ui/core/DialogContentText';
-import { useTranslation } from 'react-i18next';
-import { createPost, editSelectedPost } from '../../../modules/post';
-import {
-  useStyles,
-  ShareSettingsWrapper,
-  RespFormGroup,
-  ArticleInfo
-} from './ShareSettings.styles';
-
-export default function ShareSettings({
-  newPost,
-  resetContent,
-  edit,
-  postObj,
-  isArticle = false
-}) {
-  const history = useHistory();
-  const dispatch = useDispatch();
-  const classes = useStyles();
-  const location = useLocation();
-  const [shareState, setShareState] = useState({
-    shareWithFriends: true,
-    shareAnonymously: false
-  });
-  const [noContentDialogOpen, setNoContentDialogOpen] = useState(false);
-
-  useEffect(() => {
-    if (location.pathname === '/anonymous') {
-      setShareState({ shareWithFriends: false, shareAnonymously: true });
-    } else if (
-      location.pathname?.includes('/articles') ||
-      location.pathname?.includes('/responses')
-    ) {
-      const { share_with_friends, share_anonymously } = postObj;
-      setShareState({
-        shareWithFriends: share_with_friends,
-        shareAnonymously: share_anonymously
-      });
-    }
-    if (isArticle) {
-      setShareState({ shareWithFriends: true, shareAnonymously: false });
-    }
-  }, [isArticle, location, postObj]);
-
-  const [t] = useTranslation('translation', { keyPrefix: 'feed_common' });
-
-  const handleChange = (e) => {
-    const { name, checked } = e.target;
-    setShareState((prev) => ({ ...prev, [name]: checked }));
-  };
-
-  const onClickSubmitButton = async () => {
-    if (edit) {
-      if (postObj?.content.trim() === '') {
-        setNoContentDialogOpen(true);
-        return;
-      }
-      const editedPostObj = {
-        ...postObj,
-        share_with_friends: shareState.shareWithFriends,
-        share_anonymously: shareState.shareAnonymously
-      };
-      await dispatch(editSelectedPost(editedPostObj));
-      history.push(location.pathname.slice(0, -4));
-    } else {
-      if (newPost?.content.trim() === '') {
-        setNoContentDialogOpen(true);
-        return;
-      }
-      const newPostObj = {
-        ...shareState,
-        ...newPost
-      };
-      dispatch(createPost(newPostObj));
-      resetContent();
-      if (location.pathname === '/anonymous') {
-        setShareState({ shareWithFriends: false, shareAnonymously: true });
-      } else {
-        setShareState({ shareWithFriends: true, shareAnonymously: false });
-      }
-    }
-  };
-
-  const controlShareWithFriends = (
-    <Checkbox
-      id="share-with-friends"
-      name="shareWithFriends"
-      checked={shareState.shareWithFriends}
-      onChange={handleChange}
-      size="small"
-    />
-  );
-
-  const controlShareAnonymously = (
-    <Checkbox
-      id="share-anonymously"
-      name="shareAnonymously"
-      checked={shareState.shareAnonymously}
-      onChange={handleChange}
-      size="small"
-    />
-  );
-
-  return (
-    <ShareSettingsWrapper>
-      <RespFormGroup row width="400px">
-        {!isArticle && (
-          <>
-            <FormControlLabel
-              className={`share-with-friends ${classes.label}`}
-              control={controlShareWithFriends}
-              label={
-                <Typography className={classes.label}>
-                  {t('share_with_friends')}
-                </Typography>
-              }
-            />
-            <FormControlLabel
-              className={`share-anonymously ${classes.label}`}
-              control={controlShareAnonymously}
-              label={
-                <Typography className={classes.label}>
-                  {t('share_anonymously')}
-                </Typography>
-              }
-            />
-          </>
-        )}
-        <Button
-          id="submit-button"
-          size="small"
-          variant="outlined"
-          color="secondary"
-          onClick={onClickSubmitButton}
-          disabled={
-            (!shareState.shareAnonymously && !shareState.shareWithFriends) ||
-            (edit ? !postObj?.content : !newPost?.content)
-          }
-        >
-          {t('post')}
-        </Button>
-      </RespFormGroup>
-      {isArticle && (
-        <ArticleInfo>
-          {t('posts_excluding_questions_and_answers_are_only_open_to_friends')}
-        </ArticleInfo>
-      )}
-      <Dialog
-        id="no-content-dialog"
-        open={noContentDialogOpen}
-        onClose={() => setNoContentDialogOpen(false)}
-        aria-labelledby="alert-dialog-title"
-        aria-describedby="alert-dialog-description"
-      >
-        <DialogContent>
-          <DialogContentText id="alert-dialog-description">
-            {t('please_enter_the_content')}
-          </DialogContentText>
-        </DialogContent>
-        <DialogActions>
-          <Button
-            id="confirm-button"
-            onClick={() => setNoContentDialogOpen(false)}
-            color="primary"
-            autoFocus
-          >
-            {t('ok')}
-          </Button>
-        </DialogActions>
-      </Dialog>
-    </ShareSettingsWrapper>
-  );
-}
